Move error message classes to errors package
