+++ server/bulletin/errors/membership.py
@@ -1,0 +1,4 @@
+class MembershipErrorMessage:
+    INSUFFICIENT_PRIVILEGES = 'You have insufficient privileges to ' \
+                              'perform this action.'
+    NO_BOARD_ACCESS = 'You do not have access to this board.'
